requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
motor
python-dotenv